close-inactive/requirements.txt
mcstatus
google-cloud-monitoring
google-cloud-compute
